+++ src/type_check.h
@@ -1,11 +1,11 @@
 #ifndef TYPE_CHECK_H
 #define TYPE_CHECK_H
 
-unsigned int
+short int
 is_dir(const char* dest_name);
 
 /* I have a feeling I'll need this*/
-unsigned int
+short int
 is_existing_file(const char* file_name);
 
 #endif
